Fix get.sh for CentOS9
Signed-off-by: Richard Gee <richard@technologee.co.uk>

When running `curl -sLS https://get.arkade.dev | sudo sh` on CentOS9 the script exits with a non-zero exit code. Despite this arkade is installed.

This seems to be affecting use of the command under `sudo` as calling `arkade` outside of sudo returns successfully.  This would seem to be due to the PATH setting in `/etc/sudoers` not looking in the installed directory.

This change updates the `get.sh` variable for `SUCCESS_CMD` so be explicit about the binary's location.